Refactor: 1차 전체 수정 (#23)
* chore: 개발 초기 세팅

* feat: add BaseEntity Getter

* feat: Comment Swagger 문서화

* feat: Review Swagger 문서화

* chore: 오타수정

* feat: 댓글 외래키 인덱스 추가

* feat: 댓글 api 구현

* feat: 리뷰 인덱싱 추가

* feat: 리뷰 API 구현

* feat: lazy loading 적용

* feat: join column 명시

* feat: cursor paging 조건 추가

* chore: 패키지명 변경

* feat: 댓글간 연관관계 수정

* chore: 불필요 매개변수 제거

* fix: validation 수정

* chore: 최상위 .idea ignore

* chore: 필요없는 설정파일 제거

* chore: 패키지 구조 변경

* feat: 기본 엔티티 생성

* fix: 엔티티 정의 수정

* chore: application 설정 수정

* feat: dto 추가

* fix: 패키지 최신화

* chore: 팀 코드리뷰 변경사항 반영

---------

+++ matjalalzz/src/main/java/shop/matjalalzz/user/dto/OAuthSignUpRequest.java
@@ -6,18 +6,20 @@ import jakarta.validation.constraints.NotNull;
 import shop.matjalalzz.user.entity.enums.Gender;
 
 public record OAuthSignUpRequest(
-	@NotBlank(message = "닉네임은 필수 입력값입니다.")
-	String nickname,
+    @NotBlank(message = "닉네임은 필수 입력값입니다.")
+    String nickname,
 
-	@NotBlank(message = "전화번호는 필수 입력값입니다.")
-	String phoneNumber,
+    @NotBlank(message = "전화번호는 필수 입력값입니다.")
+    String phoneNumber,
 
-	@NotBlank(message = "이름은 필수 입력값입니다.")
-	String name,
+    @NotBlank(message = "이름은 필수 입력값입니다.")
+    String name,
 
-	@Min(value = 0, message = "나이는 0 이상이어야 합니다.")
-	long age,
+    @Min(value = 0, message = "나이는 0 이상이어야 합니다.")
+    int age,
 
-	@NotNull(message = "성별은 필수 입력값입니다.")
-	Gender gender
-) {}
+    @NotNull(message = "성별은 필수 입력값입니다.")
+    Gender gender
+) {
+
+}
